Fix a signed/unsigned comparison warning in fts5.
FossilOrigin-Name: 85eb05ddbc3a7cff986af13c3dab6c0063e62171

+++ ext/fts5/fts5_buffer.c
@@ -16,7 +16,7 @@
 #include "fts5Int.h"
 
 int sqlite3Fts5BufferSize(int *pRc, Fts5Buffer *pBuf, u32 nByte){
-  if( pBuf->nSpace<nByte ){
+  if( (u32)pBuf->nSpace<nByte ){
     u32 nNew = pBuf->nSpace ? pBuf->nSpace : 64;
     u8 *pNew;
     while( nNew<nByte ){
